requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.2.1
motor==3.3.1
pymongo==4.5.0
python-multipart==0.0.21
PyPDF2==3.0.1
python-docx==1.2.0
requests==2.32.5